Allow setting of CLOSURE_ARGS in ports (#24192)
This is the implementation that was discussed in #24109 

I can update the Changelog in a later commit if the tests pass in CI and
this PR is approved.

* I manually/locally ran `other.test_closure_externs` to make sure that
the changes were not affecting existing code
* I added a test `other.test_closure_args_from_port` that piggy back on
`test_closure_externs` to pass the same parameter in the port instead
and make sure that these changes actually do what they are supposed to
do in the first place.

Hopefully I didn't forget anything. The changes are pretty
straightforward in the grand scheme of things.