Deprecate -sUSE_WEBGPU in favor of Dawn's emdawnwebgpu
Adds a linker warning for using -sUSE_WEBGPU, and update documentation.

Emscripten's implementation of webgpu.h is unmaintained. We want people
to move to Dawn's emdawnwebgpu implementation (fork). However, the API
has changed substantially, so this not a clean transition. Plus, there
might be new bugs in that implementation. So we need a decently long
deprecation period before removing the old implementation.

Issue: 23432