[embind] Allow pointers for properties with a getter and setter function. (#24239)
When using separate getter and setter functions for class properties the
setter function did not have policies applied to it. This caused setters
with raw pointers to fail. For now, use the supplied return value policy
to allow pointers. In the future we should probably add separate polices
for arguments, but that will be a much bigger change.

Note: I've added a lot more testing for properties and return value
polices. This has highlighted a few issues where we seem to be creating
extra copies.

Fixes: #24225