Remove no-op branch from acorn-optimizer
This line is equivalent to `(!fun.type) === 'FunctionDeclaration'` which is equivalent to `false === 'FunctionDeclaration'` which is equivalent to just `if (false)`.

My first thought was to just fix the syntax to `!==` which I assume was the intent here, but @kripken introduced that line 4 years ago and that `if (...) continue` was never hit with no obvious (?) issues so far, so I checked with him and we decided removing that branch is the best choice.