Add typings for explicit resource management (#24196)
This is a follow-up to
https://github.com/emscripten-core/emscripten/pull/23818.

When targeting an existing ES version like ES2024, TypeScript transpiles
`using ...` and doesn't require any typings.

However, when targeting ESNext, it expects `using ...` to be a native
feature, and then looks up `[Symbol.dispose]` in the type and fails with
a type validation error if it doesn't exist.

Since there's no way to do conditional compilation based on target
version in TS, I'm adding this property to type declarations but marking
with `ts-ignore` so that `Symbol.dispose` not existing is not an error
in versions below ESNext.