[libbrowser.js] Remove non-standard Module exports
Anyone who wants to use these functions on the Module object can export
them using e.g. `-sEXPORTED_FUNCTIONS=getUserMedia`.

As part of this I had to inject the export of `createContext` for SDL2
and SDL3 which were using `Module['createContext']` for some reason.
I'm working on upstream patches to remove this usage.