Don't treat internal globals as expected user exports. NFC
This change renames `treat_as_user_export` to `is_user_export` and
inverts the relationship between this function and `is_internal_global`.

Internal globals is a subset of all non-user symbols, not the other
way around.

This change is split of from a larger ESM integration change I'm
working on.