Delete invalid test
Test expectations were incorrect: it mixed up int64/uint64 and was returning negative number in uint64 function and expecting the negative sign to be preserved. Now that behaviour of uint64 is fixed, this expectation doesn't hold.

Additionally, it was relying on the `n` suffix being printed by `console.log`, which is not portable and is not printed in e.g. latest Node.js.

The correct test is now better served by test_i64_binding which correctly checks the type limits, signs, and will also run across more engines.