[esm-integration] Fix handling of EM_JS exports
This change also adds a new core test mode for ESM integration and
uses this to run all the variants of test_fs_js_api.

As part of this I also updated the octal constants in test_fs_js_api.c
which are required to be of the new form in strict JS.  All code in ES
modules is implicitly in strict mode.

Split out from #24288

See #24060