Simplify and fix SAFE_HEAP (#24291)
This streamlines the `SAFE_HEAP` transform in acorn-optimizer. I was
going to send this as a separate cleanup PR as part of my other work on
acorn-optimizer, but it happens to also be the easy fix for a regression
mentioned in https://github.com/emscripten-core/emscripten/pull/24289
and caused by https://github.com/emscripten-core/emscripten/pull/24283
which added support for 64-bit integers to acorn-optimizer passes.

I could fix that issue separately by tweaking `unSign`, but this way we
don't need to bother - the values with correct sign are now
automatically retrieved by accessing the correct `HEAP*` array.

I've added a regression test for the linked issue as well, which fails
before this PR and passes after.