[embind] Initialize embind on workers after main thread ctors.
Previously, when a statically constructed thread was created it could
potentially run before all of the embind static constructors on the main
thread had run. This caused the workers to miss the embind bindings.

To fix this, I've added a `std::atomic` so that each worker waits for
embind to be ready before running the binding code.