Apply https://github.com/llvm/llvm-project/pull/133999
LLVM 20 release contains a bug that it one of the headers in
`libc/src/__support` directory includes a header from `libc/src/errno`,
which is not one of the header directories that are meant to be
included:
https://github.com/llvm/llvm-project/blob/ec28b8f9cc7f2ac187d8a617a6d08d5e56f9120e/libc/src/__support/str_to_float.h#L34

I was told
(https://github.com/llvm/llvm-project/pull/91651#issuecomment-2880876115)
this was a bug that was fixed in
https://github.com/llvm/llvm-project/pull/133999. Whether this fix will
be backported to later version of LLVM 20 release is unclear. To avoid
importing files from `libc/src`, this applies the fix here as well.