Unify Node detection code
This was duplicated in quite a few places and slightly different in some
places.

This change reduces that size of the code a little and reduces the
number of places we specify it to just two.

This change also fixes the instance.test_environment test which was
failing due to broken nodejs detection in shell.js (it was checking
for `require` which doesn't exist in ES module mode).