Simplify GROWABLE_HEAP transform (#24295)
Similar to https://github.com/emscripten-core/emscripten/pull/24291, but
for the transform handling `-pthread` + `-sALLOW_MEMORY_GROWTH`.

In the first commit I added a codesize variation to commit current
sizes, and in the second you can see that this transform actually saves
the total size in addition to making acorn-optimizer simpler.