Use AsciiToString in Embind. NFC
Embind had its own read readLatin1String, which is functionally identical to the AsciiToString helper we have in the shared libstrings library.

The only difference is that It had extra logic for caching characters by code, which are likely not even an optimisation in modern JIT engines, and if we do want to do that in the future for some reason, it's best to do so in centralised helpers.

For now, a simple search-replace gives an easy win in terms of code size.