Unify legacy dyncall mechanisms (#24371)
With this change we use the MINIMAL_RUNTIME technique in all cases.

That is, dyncall functions are stored in a global `dynCalls` object that
maps signatures to functions. Previously we were relying on the
`dynCall_xx` helpers existing on the `Module` object.

This allows `-sMODULARIZE=instance` to work with `-sDYNCALL` and by
extension with `-sASYNCIFY=1`.