Cleanup run_codesize_test. NFC
This change removes the `expected_exists`, `expected_not_exists`
arguments which are no longer used.

We had a few places where `waka` was being passed to
`expected_not_exists` but there are no code size tests that use a
symbols called `waka`.  Perhaps there once was and this was just
cargo-culted?

Also, remove the now-unnecessary use of `*args` which was making it hard
to understand exactly which arguments the tests were using.