Add full expected JS output test_codesize_minimal_O0
This file will be quite volatile since it will change on almost any
change to the code JS code we generate.  However it will make is much
easier to see the effect of changes like #24384.

Also, it should not be any more of less volatile than the existing code
size expectations files and, like those, this file is updated
automatically using the `--rebaseline` test flag.