[ASYNCIFY] Avoid dependency on the names of exports (#24382)
This change moves from using export names to function identity to track
the functions in the resumption stack.

The upside of this is that we can wrap non-exports, such as table
entire, which we currently avoid by enabling `-sDYNCALLS` whenever
`ASYNCIFY=1` is used.

It also means we can potentially make `ASYNCIFY=1` work with the
`wasmExports` global existing (i.e. in `WASM_ESM_INTEGRATION` mode.