Avoid generating JS thunks for Wasm exports (#24384)
This change unifies the way we handle Wasm exports between various
different modes.

Now, we no longer generate thunks for Wasm exports before the module is
instantiated. Instead, in debug mode we generate trapping functions that
let users know that they used a function reference that was taken too
early.

All modes (except DECLARE_ASM_MODULE_EXPORTS=0) now generate a function
called `assignWasmExports` that gets called once the module is
instantiated and update the global symbols as the Module exports.

This is code size win overall.

This change also simplifies the emitDCEGraph and applyDCEGraphRemovals
because it reduces the number of ways the wasmExports are used.