Deprecate -sUSE_WEBGPU in favor of --use-port=emdawnwebgpu (#24220)
Add a linker warning for using `-sUSE_WEBGPU` pointing to
`--use-port=emdawnwebgpu`, and update documentation.

Moves the port from `contrib.emdawnwebgpu` to just `emdawnwebgpu` since
we'll keep maintaining it as we work on Emdawnwebgpu.

> Emscripten's implementation of webgpu.h is unmaintained. Dawn now
maintains "emdawnwebgpu", a fork of Emscripten's implementation that is
up to date, resolving
[#23432](https://github.com/emscripten-core/emscripten/issues/23432).
> 
> We want people to move to emdawnwebgpu. However, the API has changed
substantially, so it will take people time to switch over. Plus, there
might be new bugs in that implementation. So we need a decently long
deprecation period before removing the old implementation.

Closes #23432
Starts fixing #24265