Fix sysroot installation when emscripten itself is read-only (#24405)
The resulting header tree need to be at least user-writable so that we
can incrementally copy files into it, adding new files and also updating
previously copied ones.

Sadly, python's shutil doesn't seem to have any options for disabling
the copying of mode bits.

This was originally fixed in #15386 but then regressed in #23417.

Fixes: #24404