Remove/fix usage of MIN_SAFARI_VERSION
We don't support targeting safari version older than 10.10.00 so two
of these conditions were always true.

The third check was added in #21428 and looks to me like it was always
wrong.  The window.orientation API was added in safari 16.4 as the
comment says so I updated the check to match.  I'm not sure where the
old value of 0x100400 came from or what it means.