[MODUALRIZE] Remove redundant try/catch in createWasm (#24418)
In `MODUALRIZE` mode we always do `await createWasm` before we return
the readyPromise.

This means there is no need to call `readyPromiseReject` inside of
createWasm.

Furthermore, we can completely avoid creating the ready promise if we
don't need it and make all the resolve/reject sites call the
resolve/reject handlers conditionally.

Fixes: #24415