Refactor gen_struct_info to emit JSON directly. NFC (#24489)
I wanted to add something to gen_struct_info functionality, but found it
difficult to comprehend the custom intermediate format of the C
program's output.

I changed it to emit the desired result directly as JSON instead, which
results in the same output but makes code easier to reason about and
debug, as well as possibly tiny bit faster as well.