Use higher-level string helpers where possible. NFC (#24486)
I guess one functional change is in implementation of
`emscripten_fetch_get_response_headers_length`.

The docs say that it just returns the length, and you must use `length +
1` for buffer size passed to `emscripten_fetch_get_response_headers`,
and that's what tests do, but we were actually returning `length + 1`
already. This seems like a bugfix, but happy to revert if deemed too
risky.