Remove libformatString.js (#24508)
After a recent change in
https://github.com/emscripten-core/emscripten/pull/24489 we no longer
need `formatString` for bootstrapping, and there is only one place
remaining that uses this helper.

As far as I could track, it was added in very early days of Emscripten
when `printf`, `fprintf` and all other friends were implemented directly
in JS, and so needed a JS-based implementation of this formatting as
well. Nowadays, we use native implementation from C, so usage of
`formatString` can be easily replaced with `vasprintf` (or any other
relevant `sprintf` variation).

After replacing it in the implementation of `emscripten_log`, the only
other function that was used from this file was `unSign` in a test code,
but that was no-op (it was unsigning items in a Uint8Array which is
already unsigned by definition), so this entire file can be removed now.

My only concern is, as usual, whether these functions can be considered
internal - they were never documented, so I hope so?