[contrib.glfw3] Use new external port feature (#24518)
Now that 4.0.10 has been released and contains the new feature to make a
port external, I went ahead and applied it to `contrib.glfw3`. In the
end I really like this model because it keeps all the implementation
details of the port outside Emscripten. Incidentally it makes
reading/parsing the port file faster and also reduces the size that gets
shipped with Emscripten, especially if you never need the port...