Fix `makeRemovedFSAssert()` resolved path (#24567)
This fixes an issue where `path.resolve()` resolves using the CWD (which
is usually the target project path) instead of the directory of the
script itself.

The current behavior [causes issues with the Closure
compiler](https://github.com/godotengine/godot/pull/107460#issuecomment-2970895256),
as stubs are erroneously created instead of being skipped in
`src/shell.js`. Closure then complains that variables are declared
twice.

The issue was caused by #23348 which changed the logic of the search.