[test] Measure gzip sizes without metadata. NFC
Writing to a temporary gzip file, measuring and then deleting is both inefficient, but also results in a slightly larger reported size because it then also writes extra metadata like the original filename.

That metadata is not something we care about measuring, as more often than not compression on the Web happens on the fly anyway.