[NFC][Testing] Test Poppler with optimizations (#24628)
Poppler's configure script chooses -O2 for gcc, and nothing for all other
compilers, so we were never optimizing source files. Fix that by adding
optimizations in cflags.

This is important for test coverage, but also for branch hints, as LLVM
does not emit hints when not optimizing. Another fix for branch hints
is applied here: the likely macro was tailored (once again) to gcc. (This
is NFC for now, but if/when LLVM gets branch hint support, this will
become a useful real-world test of compiling and linking real-world code
with such hints.)

This also makes the Poppler benchmark 33% faster.