[embind] Don't truncate `unsigned long` under wasm64 (#24678)
This makes it consistent with other bindings like EM_ASM.

As part of this, I had to change vector and map sizes and indices to
explicitly take `unsigned int` instead, so that users don't have to
upgrade all their code to deal with unexpected bigints in e.g.
`vec.size()`.

Technically this means you can't pass around a C++ vector >4GB to JS and
back, but I'm sure we can live with it.