Use the more optimal overload for copying `val` (#24694)
I found by accident (while stepping via the debugger) that for `val(v)`
where `v` is `val&` C++ chooses the `T&&` overload as more precise one
instead of `const val&`.

The runtime behaviour of the two is equivalent, so I'm not sure it can
be easily tested, but forwarding to the const overload would certainly
be a bit more optimal.