Use threaded HTTP server for test harness. (#24748)
The test `test_offset_converter_pthread` hangs during the `fetch` call
to report results to the server. This seems to be caused by Chrome
opening an earlier connection to the server and leaving it open which in
turn causes pythons`SimpleHTTPRequestHandler` to stop responding to new
requests. Using the `ThreadingHTTPServer` allows python to keep
processing requests and fixes the hang.

More info on this issue can be found in:
https://bugs.python.org/issue31639 and the chrome issue
https://issues.chromium.org/issues/40978518.

Fixes #10232