Simplify MODULARIZE code to always use async factory function (#24727)
The docs here were wrong. We have the `addRunDependency` system that can
delay module's readiness for an arbitrary amount of time, even if async
compilation is disabled.

This `addRunDependency` system does not exist in `MINUMAL_RUTTIME`, so
in some cases we can avoid the ready promise in `MINIMAL_RUNTIME` mode,
but that is a pretty subtle distinction that only applies to
`MINIMAL_RUNTIME`. Also even in this mode we were still marking the
factory function as async (note the lack of `USE_READY_PROMISE` in the
existing condition).