Avoid injecting `--experimental-wasm-bigint` into `#!` line when creating node-runnable output (#24808)
This was the last remaining usage of `node_bigint_flags` in the compiler
so this function
was moved into test code where it is still used.

IIUC its better not to do this at all for a few reasons:

1. The version of node used in the emscripten compiler is not the same
as the version that will be used the run the output program, so basing
these flags on `settings.NODE_JS` is just wrong.
2. If folks want to target older versions of node we already have ways
to do that `-sMIN_NODE_VERSION`. They can also add this flag themselves
when running the output program. I don't think is out business to inject
node flags here.

---------

Co-authored-by: Alon Zakai <alonzakai@gmail.com>