[embind] Add pointer policies for setting val object properties. (#24811)
After #24175, `val` required a pointer policy when creating a `val` with
a pointer. In `val.set(key, value)`, a temporary `val` object is
created, but there was no way to set the policy from user code. This
patch allows the user to pass in `allow_raw_pointers()` to enable this.