Fix sanitizer builds of test_minimal_runtime_hello_world_streaming. NFC
This test updated in #24849 which caused the asan and lsan versions of
the test to start failing.  The reason is that prior to #24849 there was
a bug that caused the test not actually to be run via the santizers
since self.cflags was being assigned using `=` rather then `+=`.  This
was clobbering all the builtin cflags such as `-fsanitizer=address`.

Once #24849 this bug was fixed but that exposed another bug in the test
which is that it was not including the
`minimal_runtime_exit_handling.js` file.  Without this file included
tests that use MINIMAL_RUNTIME and EXIT_RUNTIME would see an unhandled
excpetion then the program exits (due to the `throw` of `exit(0)`).