[AUDIO_WORKLET] fix emscripten_create_audio_context for wasm64 (#24866)
I spotted there was no coverage for `emscripten_create_audio_context()`
and then that it fails with wasm64 (it accesses the heap directly). This
is a minor fix with an addition to the test.

Note: the code is slightly smaller than the original but still performs
the same (omitting `latencyHint` or `sampleRate` results in
`undefined`).