Fix WASM_WORKERS + MINIMAL_RUNTIME to transfer compiled module (#24865)
The current code was transferring `Module['wasm']` which, on them the
main thread is normally a byte array and not the compiled module. The
new code uses the same pattern that pthreads uses to stash the compiled
module and postMessage that instead, which should be more performant.

In the worker / pthread the global `Module['wasm']` is then set to the
received object so somewhat confusingly `Module['wasm']` is always a
compiled module in a worker / pthread (at least under
`MINIMAL_RUNTIME`).