Remove internal USE_READY_PROMISE (#24874)
This was only used in one very specific case which required all the
following to be enabled:

- MINIMAL_RUNTIME
- MODULARIZE
- SINGLE_FILE
- !PTHREADS
- html output

This combination rare because most folks don't have use the emscripten
generated html output. So that rules out most of our users, then they
would also have to choose `MODULARIZE` but not actually want to create
multiple instances or pass in any arguments. This is because in order to
support this very special use case the generated HTML was setup to stamp
out a singleton instance of the module with no arguments passed in.

While it *is* sometimes useful to stamp out a single instance of the
module we don't do this in any other circumstances. If a user wants to
do this its easy enough to do with a `--post-js` file. In fact we do
this in our test code when we want to test the `-sMODULARIZE` setting.
See `test/modularize_post_js.js`.

So think this change is really an improvment and I think its very
unlikely that many folks will be using this specific combination of
settings.