[AUDIO_WORKLET] Tidy and document the worklet processing code. NFC (#24878)
This renames some of the single-letter variables and documents their
use.

Since I've spent a lot of time in this code it's extra work to keep
remembering that `k` is the struct address (previously the heap index),
`j` is the audio channel, etc. I also removed some of the code that had
side effects, e.g. `c = (a += b)`; some but not all).

The end result (IMO) is readable and even saves a few bytes after
Closure.

(This is the only planned change like this ahead of the optimisation PR)