Roll emdawnwebgpu port to v20250807
From https://github.com/google/dawn/releases/tag/v20250807.221415

This version fixes the known showstopping bugs. It also improves the
documentation included in the port file so it's easier to understand
what to do with it.

After this has landed for a little while it will be a good time to
finally remove -sUSE_WEBGPU from Emscripten. (#24265)