[file_packager] Remove unused `errback` fetchRemovePackage. NFC
This argument has basically used since 3839dd4de.

Even when this function was first add in 74f3242118d the errback
callback was only used for the IDB failures, not the fetch failures.
Fetch failure were left to propagate to the global error handlers.

The exception here was that when node support was added in #11785 it,
I believe incorrectly used errback to handle file read errors (doesn't
match the web version that does *not* handle fetch error using errback).

This is small cleanup in advance of #24893.