[file_packager] Remove unused `errback` fetchRemovePackage. NFC
This argument has been used since 3839dd4de.

Even when this function was first add in 74f3242118d the errback
callback was only used for the IDB failures, not the fetch failures.
Fetch failure were left to propagate to the global error handlers.

This is small cleanup in advance of #24893.