Convert `FS_handledByPreloadPlugin` and plugin `handle` callback to async. (#24914)
The `FS_handledByPreloadPlugin` is completely internal so changing this
should be completely non-functional.

If external projects are implementing custom preload plugins they would
need to update their `handle` methods.  I've added an assertion in debug
builds to detect plugins that are not async.

This change will allow a followup which is to convert that whole file
preloading
mechanism to async.