Fix regression to enable specifying EMTEST_BROWSER env. var in native path format on Windows. (#24962)
Fix regression to enable specifying EMTEST_BROWSER env. var in native
path format on Windows.

This is something that used to work a long time ago, and at some
refactor got lost. On Windows environment variables that point to
directories or files canonically use native path slashes `\` to do so.