Fix core3.test_asyncify_lists* when Emscripten SDK was built from source. (#24997)
The `self.assertFalse(b'main' in binary)` check would fail because the
binary does, and should, contain the name `main`, in an assert() check.
In disassembly:

```
 (data $0 (i32.const 1024) "c_baz\00-+   0X0x\00c:\\emsdk\\emscripten\\main\\test\\core\\test_asyncify_lists.cpp\00foo\00a == b\00(null)\00bar: %d\n\00foo: %d\n\00c++: %d\n")
```

i.e. the occurrence of `main` is in a path name.

Change the code to check another symbol, and verify that the symbol name
is caught in some other mode so it won't become tautologically true in a
future refactor.