Fix race condition in checkMailbox
The `checkMailbox` callback can occur after the thread has terminated.
In this case calling into native code can trigger the `makeAbortWrapper`
wrapper that is put around each native function which then results in a
"program has already aborted!" error being thrown.

Once solution to this is to make sure that the function which are called
do not have `makeAbortWrapper` applied to them.

This was the technique I used in #18754, but the list of functions
became stale when emscripten_proxy_execute_task_queue was removed in #18852.

A better solution is to wrap to whole function in callUserCallback,
which takes case of checking if the runtime is alive before calling into
native code.

Fixes: #20067