When printing test harness stuff, print to stderr to be more consistent. (#25081)
Iirc the original rationale for the test harness was that test harness
infra related prints would go to stderr, and tests itself would print to
stdout? (at least that is how I gathered it to be from the start of
time..)

Most of the harness stuff goes to stderr, though a few stray lines do
not. So move those to print to stderr also to be more consistent.

When observing the prints on Buildbot, this distinction helps a bit
visually, since stderr is printed red, and individual test output then
in gray.