Add script to split module based on source paths
This adds a script, `tools/empath-split.py`, which is a wrapper for
Binaryen's `wasm-split`. `wasm-split` has `--multi-split` mode, which
takes a manifest file that lists the name of functions per module.
(Example:
 https://github.com/WebAssembly/binaryen/blob/main/test/lit/wasm-split/multi-split.wast.manifest)

But listing all functions belonging to each module is a tedious process.
`empath-split` takes a wasm file and a text file that has a list of
paths, which can be either directories or functions, and using the
source map information, generates a manifest file, and runs
`wasm-split`.

This makes a small drive-by fix for `emsymbolizer`. Currently when it
takes a 0 address, it returns the location info associated with
offsets[-1], which is the largest offset. This fixes it, and adds an
optional `lower_bound` argument to `find_offset` so that when we want to
get a source info entry, we don't go below the current function start
offset.